(PC-17655) chore(Accessiblity): move tests next to their component
